requirements.txt
keras
numpy
tensorflow
matplotlib
opencv-python
lime
huggingface_hub
datetime
scikit-image
pillow
transformers
tqdm
safetensors
torch
torchvision
